+++ packages/ui-components/web/src/Modals/SearchInputModal.tsx
@@ -4,6 +4,7 @@ import { RichContentTheme } from 'wix-rich-content-common';
 import { TextSearchInput } from '../';
 import { KEYS_CHARCODE } from '../consts';
 import ModalContainer from './TextInputModalContainer';
+import styles from '../../statics/styles/url-input-modal.scss';
 
 const SearchInputModal = props => {
   interface ModalProps {
@@ -40,7 +41,7 @@ const SearchInputModal = props => {
 
   return (
     <ModalContainer {...props}>
-      <>
+      <div className={styles.textSearchInput_container}>
         <TextSearchInput
           id="dropdown-text-input"
           onClose={onCloseRequested}
@@ -53,7 +54,7 @@ const SearchInputModal = props => {
           autoComplete="off"
         />
         {children}
-      </>
+      </div>
     </ModalContainer>
   );
 };