+++ CHANGELOG.md
@@ -16,15 +16,40 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+  
+### :bug: Bug Fix
 
+- `ricos-content`
+  - [#3019](https://github.com/wix/ricos/pull/3019) fromDraft: Invalid entity key protected
 
 ### :nail_care: Polish
 
+- `ui-components`
+  - [#2999](https://github.com/wix/ricos/pull/2999) fix action buttons keyboard navigation order
+- `vertical-embed`
+  - [#3005](https://github.com/wix/ricos/pull/3005) vertical-embed modal new design
 - `line-spacing`
+
   - [#3009](https://github.com/wix/ricos/pull/3009) add custom line-height selection to the line-heights options display
+- `ricos-content`
+  - [#3016](https://github.com/wix/ricos/pull/3016) Polls Content Builder API support
+
+### :house: Internal
+
+- `schema`
+  - [#3013](https://github.com/wix/ricos/pull/3013) undefined fields added to plugin defaults JSON
+- `divider`
+  - [#3017](https://github.com/wix/ricos/pull/3017) set divider width to 100% and remove editor bounds consideration
 
 </details>
 <hr/>
+
+## 8.56.5 (Oct 18, 2021)
+
+### :house: Internal
+
+- `ricos-content`
+  - [#96dce6a](https://github.com/wix/ricos/commit/96dce6a93b8f4227dfa86c801b77b349049a4ee0)velo adapter - image element source's regex fix
 
 ## 8.56.4 (Oct 18, 2021)